# frontend/swagger.json
{
    "openapi": "3.0.0",
    "paths": {
        "/auth/login": {
            "post": {
                "operationId": "AuthController_login",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "operationId": "AuthController_register",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RegisterDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                }
            }
        },
        "/profile": {
            "get": {
                "operationId": "AuthController_profile",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/products": {
            "get": {
                "operationId": "ProductsController_list",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            },
            "post": {
                "operationId": "ProductsController_create",
                "parameters": [],
                "responses": {
                    "201": {
                        "description": ""
                    }
                }
            }
        },
        "/products/{id}": {
            "get": {
                "operationId": "ProductsController_detail",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            },
            "put": {
                "operationId": "ProductsController_update",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            },
            "delete": {
                "operationId": "ProductsController_remove",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/categories": {
            "get": {
                "operationId": "CategoriesController_list",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            },
            "post": {
                "operationId": "CategoriesController_create",
                "parameters": [],
                "responses": {
                    "201": {
                        "description": ""
                    }
                }
            }
        },
        "/categories/{id}": {
            "put": {
                "operationId": "CategoriesController_update",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            },
            "delete": {
                "operationId": "CategoriesController_remove",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/option-groups": {
            "post": {
                "operationId": "OptionGroupsController_create",
                "parameters": [],
                "responses": {
                    "201": {
                        "description": ""
                    }
                }
            }
        },
        "/option-groups/{id}": {
            "put": {
                "operationId": "OptionGroupsController_update",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            },
            "delete": {
                "operationId": "OptionGroupsController_remove",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/options": {
            "post": {
                "operationId": "OptionsController_create",
                "parameters": [],
                "responses": {
                    "201": {
                        "description": ""
                    }
                }
            }
        },
        "/options/{id}": {
            "put": {
                "operationId": "OptionsController_update",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            },
            "delete": {
                "operationId": "OptionsController_remove",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/cart": {
            "get": {
                "operationId": "CartController_getCart",
                "parameters": [
                    {
                        "name": "anonymousId",
                        "required": true,
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/cart/merge": {
            "post": {
                "operationId": "CartController_merge",
                "parameters": [],
                "responses": {
                    "201": {
                        "description": ""
                    }
                }
            }
        },
        "/cart/items": {
            "post": {
                "operationId": "CartController_addItem",
                "parameters": [],
                "responses": {
                    "201": {
                        "description": ""
                    }
                }
            }
        },
        "/cart/items/{id}": {
            "put": {
                "operationId": "CartController_updateItem",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            },
            "delete": {
                "operationId": "CartController_removeItem",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/orders": {
            "get": {
                "operationId": "OrdersController_list",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            },
            "post": {
                "operationId": "OrdersController_create",
                "parameters": [],
                "responses": {
                    "201": {
                        "description": ""
                    }
                }
            }
        },
        "/orders/customer-orders": {
            "get": {
                "operationId": "OrdersController_listCustomerOrders",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/orders/{orderId}/status": {
            "patch": {
                "operationId": "OrdersController_updateStatus",
                "parameters": [
                    {
                        "name": "orderId",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/search/products": {
            "get": {
                "operationId": "SearchController_searchProducts",
                "parameters": [
                    {
                        "name": "q",
                        "required": true,
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/admin/users": {
            "get": {
                "operationId": "UsersController_list",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/users/{userId}": {
            "get": {
                "operationId": "UsersController_getUser",
                "parameters": [
                    {
                        "name": "userId",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/payments/checkout-session": {
            "post": {
                "operationId": "PaymentController_createCheckoutSession",
                "parameters": [],
                "responses": {
                    "201": {
                        "description": ""
                    }
                }
            }
        },
        "/payments/webhook": {
            "post": {
                "operationId": "PaymentController_handleWebhook",
                "parameters": [
                    {
                        "name": "stripe-signature",
                        "required": true,
                        "in": "header",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": ""
                    }
                }
            }
        },
        "/healthz": {
            "get": {
                "operationId": "HealthController_check",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        }
    },
    "info": {
        "title": "Floria API Gateway",
        "description": "The API description for Floria e-commerce platform",
        "version": "1.0",
        "contact": {}
    },
    "tags": [],
    "servers": [],
    "components": {
        "securitySchemes": {
            "bearer": {
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "type": "http"
            }
        },
        "schemas": {
            "LoginDto": {
                "type": "object",
                "properties": {}
            },
            "RegisterDto": {
                "type": "object",
                "properties": {}
            }
        }
    }
}